Turn Logic entegrated. Also Some early playtesting purpoesly gameplay demo(?)
Tile generetad perfectly. Also Display tiles which is UI of that tile works perfetly. Display Tile is being used because If try to move the interactable tile then there will be no interactable tile left when it finished move. So, For that problem, Display Tiles are just reperesteaion of that tile. They have no value to the scripts. They just show Player to the tile and where Player interact.